x[d][i] != 0 or 1
cplex在算x[d][i]时，得到的结果可能不是0，也不是1，而是0.0000001这样的数，所以直接判断x[d][i]>0可能会把0.0000001这样的结果当成有流量

+++ Voting5/flow.h
@@ -213,7 +213,7 @@ double judge_sum_function(VGraph &g, vector<Flow*> &flowL, int winner)
 		int src,dst;
 		double latencyTemp;
 		src=g.incL[i]->src;dst=g.incL[i]->dst;
-		latencyTemp = 1 + flowL[winner]->adj[src][dst]/(1+ g.incL[i]->capacity - flowL[winner]->adj[src][dst]);
+		latencyTemp = 1 + (double)flowL[winner]->adj[src][dst]/(1+ g.incL[i]->capacity - flowL[winner]->adj[src][dst]);
 		latencyVoting += latencyTemp * flowL[winner]->adj[src][dst];
 	}
 	return latencyVoting;
